# additional_tools/smia_kb/src/swagger_server/swagger/swagger.yaml
openapi: 3.0.4
info:
  title: SMIA Knowledge Base (KB) | HTTP/REST | API Collection
  description: |-
    <img src="https://raw.githubusercontent.com/ekhurtado/SMIA/refs/heads/smia_kb_development/additional_resources/smia_kb/src/SMIA_KB_logo.png" alt="SMIA KB logo" height="100" />
    
    This is a sample Self-configurable Manufacturing Industrial Agents (SMIA) Knowledge Base (KB) based on the OpenAPI 3.0 specification.  You can find out more about at [GitHub](https://github.com/ekhurtado/SMIA). The SMIA KB has been developed as a SMIA-compliant infrastructure service, but it is designed to be interoperable with other approaches, taking advantage of following the OpenAPI specification. It addresses the realization of a Knowledge Base based on Capability-Skill-Service (CSS) model and driven by an ontological implementation, capable of providing CSS ontology information management.

    Some useful links:
    - [SMIA GitHub repository](https://github.com/ekhurtado/SMIA)
    - [SMIA ReadTheDocs documentation project](https://smia.readthedocs.io/en/latest/)
    - [SMIA DockerHub repository](hub.docker.com/r/ekhurtado/smia/)
  termsOfService: https://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
- url: https://smiakb.gcis.ehu/api/v3
...
paths:
  /capabilities:
    get:
      tags:
      - Capability API
      summary: Returns all capabilities related to the SMIA-CSS model.
      description: Returns all capabilities related to the SMIA-CSS model. Capabilities
        are extracted from the AAS repository or added by the user through the SMIA
        KB API.
      operationId: get_all_capabilities
      responses:
        "200":
          description: Requested SMIA-CSS capabilities
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Capability"
                x-content-type: application/json
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Capability"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-openapi-router-controller: swagger_server.controllers.capability_api_controller
    post:
      tags:
      - Capability API
      summary: Add a new Capability to the SMIA KB.
      description: Add a new Capability to the SMIA KB.
      operationId: post_capability
      requestBody:
        description: SMIA-CSS Capability object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Capability"
          application/xml:
            schema:
              $ref: "#/components/schemas/Capability"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Capability"
        required: true
      responses:
        "200":
          description: SMIA-CSS Capability created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capability"
            application/xml:
              schema:
                $ref: "#/components/schemas/Capability"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

      x-openapi-router-controller: swagger_server.controllers.capability_api_controller
  /capabilities/$identifiers:
    get:
      tags:
      - Capability API
      summary: Returns all capabilities identifiers related to the SMIA-CSS model.
      description: Returns all capabilities identifiers related to the SMIA-CSS model.
        Capabilities are extracted from the AAS repository or added by the user through
        the SMIA KB API.
      operationId: get_all_capabilities_identifiers
      responses:
        "200":
          description: Requested SMIA-CSS capabilities identifiers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReferenceIRI"
                x-content-type: application/json
            application/xml:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReferenceIRI"
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-openapi-router-controller: swagger_server.controllers.capability_api_controller
  /capabilities/{capabilityIdentifier}:
    get:
      tags:
      - Capability API
      summary: Returns a specific capability related to the SMIA-CSS model.
      description: Returns a specific capability related to the SMIA-CSS model. Capabilities
        are extracted from the AAS repository or added by the user through the SMIA
        KB API.
      operationId: get_capability_by_id
      parameters:
      - name: capabilityIdentifier
        in: path
        description: The Capability's unique identifier (must be added base64-url
          encoded. The Serialization API can be used to obtain it)
        required: true
        style: simple
        explode: false
        schema:
          $ref: "#/components/schemas/ReferenceIRIencoded"
      responses:
        "200":
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Capability"
            application/xml:
              schema:
                $ref: "#/components/schemas/Capability"
        "400":
          description: Invalid ontological IRI supplied
        "404":
          description: Capability not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-openapi-router-controller: swagger_server.controllers.capability_api_controller
    put:
      tags:
      - Capability API
      summary: Updates an existing capability related to the SMIA-CSS model.
      description: Updates an existing capability related to the SMIA-CSS model. Capabilities
        are extracted from the AAS repository or added by the user through the SMIA
        KB API.
      operationId: put_capability_by_id
      parameters:
      - name: capabilityIdentifier
        in: path
        description: The Capability's unique identifier (must be added base64-url
          encoded. The Serialization API can be used to obtain it)
        required: true
        style: simple
        explode: false
        schema:
          $ref: "#/components/schemas/ReferenceIRIencoded"
      requestBody:
        description: SMIA-CSS Capability object
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Capability"
          application/xml:
            schema:
              $ref: "#/components/schemas/Capability"
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/Capability"
        required: true
      responses:
        "200":
          description: SMIA-CSS Capability updated successfully
        "400":
          description: bad request
        "404":
          description: Capability not found
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

      x-openapi-router-controller: swagger_server.controllers.capability_api_controller
    delete:
      tags:
      - Capability API
      summary: Deletes a capability related to the SMIA-CSS model.
      description: Deletes a capability related to the SMIA-CSS model.
      operationId: delete_capability_by_id
      parameters:
      - name: api_key
        in: header
        description: ""
        required: false
        style: simple
        explode: false
        schema:
          type: string
      - name: capabilityIdentifier
        in: path
        description: The Capability's unique identifier (must be added base64-url
          encoded. The Serialization API can be used to obtain it)
        required: true
        style: simple
        explode: false
        schema:
          $ref: "#/components/schemas/ReferenceIRIencoded"
      responses:
        "200":
          description: SMIA-CSS Capability deleted successfully
        "400":
          description: Invalid capability object
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

      x-openapi-router-controller: swagger_server.controllers.capability_api_controller
...
components:
  schemas:
    Error:
      required:
      - code
      - message
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    ReferenceIRI:
      type: string
      example: http://name.org/css-smia#Capability01
    ReferenceIRIencoded:
      type: string
      example: aHR0cDovL25hbWUub3JnL2Nzcy1zbWlhI0NhcGFiaWxpdHkwMQ==
...
    Capability:
      required:
      - assets
      - category
      - hasLifecycle
      - iri
      - isRealizedBy
      - name
      type: object
      properties:
        iri:
          $ref: "#/components/schemas/ReferenceIRI"
        name:
          type: string
          example: capability01
        category:
          type: string
          example: AssetCapability
          enum:
          - AgentCapability
          - AssetCapability
        hasLifecycle:
          type: string
          description: The lifecycle of the capability.
          example: ASSURANCE
          enum:
          - ASSURANCE
          - OFFER
          - REQUIREMENT
        isRealizedBy:
          minItems: 1
          type: array
          description: By which skills is the capability realized.
          xml:
            wrapped: true
          items:
            $ref: "#/components/schemas/ReferenceIRI"
        assets:
          type: array
          description: Which assets can perform this capability.
          xml:
            wrapped: true
          items:
            $ref: "#/components/schemas/Asset"
        isRestrictedBy:
          type: array
          description: By which constraints is the capability restricted.
          xml:
            wrapped: true
          items:
            $ref: "#/components/schemas/CapabilityConstraint"
      example:
        iri: http://name.org/css-smia#Capability01
        assets:
        - kind: Type
          id: http://example.com/ids/asset001
          type: http://example.com/ids/aasElement001
        - kind: Type
          id: http://example.com/ids/asset001
          type: http://example.com/ids/aasElement001
        isRestrictedBy:
        - iri: null
          name: capabilityConstraint01
          hasCondition: INVARIANT
        - iri: null
          name: capabilityConstraint01
          hasCondition: INVARIANT
        name: capability01
        isRealizedBy:
        - null
        - null
        category: AssetCapability
        hasLifecycle: ASSURANCE
      xml:
        name: capability
...
  securitySchemes:
    api_key:  # Esto se puede usar para añadir la necesidad de usar una api_key. Despues en authorization_controller se podria añadir la logica para validarla
      type: apiKey
      name: api_key
      in: header
      x-apikeyInfoFunc: swagger_server.controllers.authorization_controller.check_api_key